+++ HISTORY.md
@@ -57,6 +57,7 @@
 * adjusted Dutch translation for Boolean ops
 * adjusted Catalan translation for Boolean ops
 * adjusted Brazilian Portuguese translation for Boolean ops
+* threads: accept Boolean values in "with input list" variant of variadic Boolean ops
 
 ### 2023-02-15
 * pushed dev version to 8.2